# src/web_module/controller/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Python API doc",
    "description": "API for python POC",
    "contact": {
      "name": "Lekhraj Dinkar",
      "email": "LekhrajDinkarus@gmail.com"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/items/{item_id}": {
      "post": {
        "summary": "Handle Params",
        "operationId": "handle_params_items__item_id__post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The ID of the item (required)",
              "title": "Item Id"
            },
            "description": "The ID of the item (required)"
          },
          {
            "name": "q1",
            "in": "query",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Query string parameter (required str)",
              "title": "Q1"
            },
            "description": "Query string parameter (required str)"
          },
          {
            "name": "q2",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Query string parameter (optional int)",
              "title": "Q2"
            },
            "description": "Query string parameter (optional int)"
          },
          {
            "name": "h1",
            "in": "header",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "header (required str)",
              "title": "H1"
            },
            "description": "header (required str)"
          },
          {
            "name": "h2",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "header (optional int)",
              "title": "H2"
            },
            "description": "header (optional int)"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Request body as a dictionary",
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}